refactor
- https://github.com/handsontable/handsontable/pull/7839#discussion_r619483001
- https://github.com/handsontable/handsontable/pull/7839#discussion_r619483624
